# src: optimizations to APyFixedArray::matmul()