… lib/test/apyfixedarray/test_matmul.py
… src/apyfixedarray_wrapper.cc