### Benchmark:
```python
from apytypes import APyFixedArray

a = APyFixedArray([range(100), range(100), range(100)], bits=32, int_bits=16)
b = a.T

for _ in range(10000):
    c = b @ a
```

### Before:
```
❯ time python3 bench_matmul.py
python3 bench_matmul.py  22,76s user 0,00s system 99% cpu 22,766 total
```

### After:
```
❯ time python3 bench_matmul.py
python3 bench_matmul.py  2,78s user 0,00s system 99% cpu 2,787 total
```